requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
nfl_data_py>=0.3.2
SQLAlchemy>=2.0.29
psycopg[binary]>=3.1.12
python-dotenv>=1.0.1
typer[all]>=0.12.3
alembic>=1.13.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
